chore: improve startup time of pdl python

- lazy load litellm
- granite_io startup optimizations by pulling the class into its own file
- lazy load jsonschema
