+++ src/pdl/pdl_llms.py
@@ -1,19 +1,14 @@
+# pylint: disable=import-outside-toplevel
 import asyncio
-import os
 import threading
 from concurrent.futures import Future
-from typing import Any, Callable, Generator, Optional, TypeVar
+from typing import Any, Callable, Generator, TypeVar
 
 import httpx
-import litellm
 from dotenv import load_dotenv
-from granite_io import make_backend, make_io_processor
-from granite_io.types import ChatCompletionInputs
-from litellm import acompletion, completion
 
 from .pdl_ast import (
     ErrorBlock,
-    GraniteioModelBlock,
     LazyMessage,
     LitellmModelBlock,
     ModelInput,
@@ -25,14 +20,6 @@ from .pdl_utils import remove_none_values_from_message
 
 # Load environment variables
 load_dotenv()
-
-# If the environment has a configured OpenTelemetry exporter, tell LiteLLM
-# to do OpenTelemetry callbacks for that exporter.  Note that this may
-# require optional OpenTelemetry Python libraries that are not pyproject.toml,
-# typically opentelemetry-api, opentelemetry-sdk,
-# opentelemetry-exporter-otlp-proto-http, and opentelemetry-exporter-otlp-proto-grpc
-if os.getenv("OTEL_EXPORTER") and os.getenv("OTEL_ENDPOINT"):
-    litellm.callbacks = ["otel"]
 
 
 def _start_background_loop(loop):
@@ -61,7 +48,11 @@ class LitellmModel:
             spec = block.spec
             parameters = set_structured_decoding_parameters(spec, parameters)
             if parameters.get("mock_response") is not None:
+                import litellm
+
                 litellm.suppress_debug_info = True
+            from litellm import acompletion
+
             response = await acompletion(
                 model=model_id, messages=list(messages), stream=False, **parameters
             )
@@ -119,6 +110,8 @@ class LitellmModel:
         parameters: dict[str, Any],
     ) -> Generator[dict[str, Any], Any, Any]:
         parameters = set_structured_decoding_parameters(spec, parameters)
+        from litellm import completion
+
         response = completion(
             model=model_id,
             messages=list(messages),
@@ -133,119 +126,6 @@ class LitellmModel:
                 msg.role = "assistant"
             yield remove_none_values_from_message(msg.model_dump())
         return result
-
-
-class GraniteioModel:
-    @staticmethod
-    def processor_of_block(block: GraniteioModelBlock):
-        assert isinstance(
-            block.model, str
-        ), f"The model should be a string: {block.model}"
-        assert isinstance(
-            block.backend, (dict, str)
-        ), f"The backend should be a string or a dictionnary: {block.backend}"
-        match block.backend:
-            case {"transformers": device}:
-                assert isinstance(block.backend, dict)
-                backend = make_backend(
-                    "transformers",
-                    {
-                        "model_name": block.model,
-                        "device": device,
-                    },
-                )
-            case backend_name if isinstance(backend_name, str):
-                backend = make_backend(
-                    backend_name,
-                    {
-                        "model_name": block.model,
-                    },
-                )
-            case _:
-                assert False, f"Unexpected backend: {block.backend}"
-        processor_name = block.processor
-        if processor_name is None:
-            processor_name = block.model
-        assert isinstance(
-            processor_name, str
-        ), f"The processor should be a string: {processor_name}"
-        io_processor = make_io_processor(processor_name, backend=backend)
-        return io_processor
-
-    @staticmethod
-    def build_message(
-        messages: ModelInput,
-        parameters: Optional[dict[str, Any]],
-    ) -> ChatCompletionInputs:
-        if parameters is None:
-            parameters = {}
-        inputs = {"messages": messages} | parameters
-        return ChatCompletionInputs.model_validate(inputs)
-
-    @staticmethod
-    async def async_generate_text(
-        block: GraniteioModelBlock,
-        messages: ModelInput,
-    ) -> tuple[dict[str, Any], Any]:
-        try:
-            assert block.parameters is None or isinstance(block.parameters, dict)
-            io_processor = GraniteioModel.processor_of_block(block)
-            inputs = GraniteioModel.build_message(messages, block.parameters)
-            result = io_processor.create_chat_completion(inputs)  # pyright: ignore
-            message = result.next_message.model_dump()
-            raw_result = result.model_dump()
-            return (
-                message,
-                raw_result,
-            )
-        except Exception as exc:
-            message = f"Error during '{block.model}' model call: {repr(exc)}"
-            loc = block.location
-            raise PDLRuntimeError(
-                message,
-                loc=loc,
-                trace=ErrorBlock(msg=message, location=loc, program=block),
-            ) from exc
-
-    @staticmethod
-    def generate_text(
-        block: GraniteioModelBlock,
-        messages: ModelInput,
-    ) -> tuple[LazyMessage, PdlLazy[Any]]:
-        future = asyncio.run_coroutine_threadsafe(
-            GraniteioModel.async_generate_text(
-                block,
-                messages,
-            ),
-            _LOOP,
-        )
-        pdl_future: PdlLazy[tuple[dict[str, Any], Any]] = PdlConst(future)
-        message = lazy_apply((lambda x: x[0]), pdl_future)
-        response = lazy_apply((lambda x: x[1]), pdl_future)
-        return message, response
-
-    # @staticmethod
-    # def generate_text_stream(
-    #     model_id: str,
-    #     messages: ModelInput,
-    #     spec: Any,
-    #     parameters: dict[str, Any],
-    # ) -> Generator[dict[str, Any], Any, Any]:
-    #     parameters = set_structured_decoding_parameters(spec, parameters)
-    #     response = completion(
-    #         model=model_id,
-    #         messages=list(messages),
-    #         stream=True,
-    #         **parameters,
-    #     )
-    #     result = []
-    #     for chunk in response:
-    #         result.append(chunk.json())  # pyright: ignore
-    #         msg = chunk.choices[0].delta  # pyright: ignore
-    #         if msg.role is None:
-    #             msg.role = "assistant"
-    #         yield remove_none_values_from_message(msg.model_dump())
-    #     return result
 
 
 MapInputT = TypeVar("MapInputT")
